restructured package significantly
