+++ .github/workflows/python-app.yml
@@ -36,4 +36,4 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Run unit tests
       run: |
-        python -m unittest src/test_unittests.py
+        python -m unittest discover BernoulliDiffusion/tests/
