+++ BernoulliDiffusion/model.py
@@ -5,6 +5,8 @@ import torch.optim as optim
 from numpy.random import default_rng
 import numpy as np
 
+from BernoulliDiffusion.utils.math_utils import kl_div, entropy
+
 # KL Divergence of multivariate bernoulli distributions:
 # https://math.stackexchange.com/questions/2604566/kl-divergence-between-two-multivariate-bernoulli-distribution
 
@@ -35,7 +37,7 @@ class ReverseModel(nn.Module):
         return x
 
 
-class BernoulliDiffusion(nn.Module):
+class BernoulliDiffusionModel(nn.Module):
     def __init__(self, model, sequence_length, num_sample_steps, T):
         super().__init__()
 
@@ -46,8 +48,8 @@ class BernoulliDiffusion(nn.Module):
 
         self.T = T
 
+        # calculate the array of beta_tilde_t values from t=1 to t=T
         self.beta_tilde_t = [torch.zeros((self.sequence_length)).to(device)]
-        # self.beta_tilde_t = [torch.ones((self.sequence_length)).to(device)]
         for cur_t in range(1, self.T+1):
             self.beta_tilde_t.append((self.beta_tilde_t[cur_t-1] + self.beta_t(cur_t) - (self.beta_t(cur_t)*self.beta_tilde_t[cur_t-1])))
         self.beta_tilde_t = torch.stack(self.beta_tilde_t)
@@ -71,6 +73,7 @@ class BernoulliDiffusion(nn.Module):
         return x
 
     def beta_t(self, t):
+        '''The bit flip probability at step t of the forward process'''
         return 1.0/(self.T-t+1)
 
     def q_conditional_prob(self, x_t, t):
@@ -94,18 +97,13 @@ class BernoulliDiffusion(nn.Module):
         '''Returns a sample x_t given input x_0'''
         return torch.bernoulli(self.q_conditional_prob_wrt_x_0(x_0, t))
 
-    def kl_div(self, q, p):
-        q_clamp = torch.clamp(q, min=0.000001, max=0.999999)
-        p_clamp = torch.clamp(p, min=0.000001, max=0.999999)
-        return -1.0*torch.sum(q_clamp * torch.log(q_clamp/p_clamp) + (1-q_clamp) * torch.log((1.0-q_clamp)/(1.0-p_clamp)), dim=1)
-
     def entropy_terms(self, x_0):
         q_start = self.q_conditional_prob_wrt_x_0(x_0, 1)
         q_end = self.q_conditional_prob_wrt_x_0(x_0, self.T)
 
-        H_start = torch.sum((-1.0*q_start*torch.log2(q_start)) - ((1.0-q_start)*torch.log2(1.0-q_start)), dim=1)
-        H_end = torch.sum((-1.0*q_end*torch.log2(q_end)) - ((1.0-q_end)*torch.log2(1.0-q_end)), dim=1)
-        H_prior = (-1.0*0.5*np.log2(0.5)) - ((1.0-0.5)*np.log2(1.0-0.5))
+        H_start = entropy(q_start)
+        H_end = entropy(q_end)
+        H_prior = entropy(torch.ones_like(q_start)*0.5)
 
         return H_start, H_end, H_prior
     
@@ -119,8 +117,8 @@ class BernoulliDiffusion(nn.Module):
             beta_t = self.beta_t(t)
             left_term = x_0*(1-self.beta_tilde_t[t-1]) + 0.5*self.beta_tilde_t[t-1]
             left_term *= x_t * (1-0.5*beta_t) + (1 - x_t) * (1.5*beta_t)
-            kl_divergence = self.kl_div(left_term,
-                                        self.p_conditional_prob(x_t, t))
+            kl_divergence = kl_div(left_term,
+                                   self.p_conditional_prob(x_t, t))
             
             H_start, H_end, H_prior = self.entropy_terms(x_0)
