+++ BernoulliDiffusion/train.py
@@ -2,15 +2,15 @@ import torch
 import torch.optim as optim
 from torch.distributions import Binomial
 
-from config import Config
-from model import BernoulliDiffusion
-from data import generate_batch
+from BernoulliDiffusion.config import Config
+from BernoulliDiffusion.model import BernoulliDiffusionModel
+from BernoulliDiffusion.data import generate_batch
 
 device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
 
 
 def train_diffusion_model(cfg: Config,
-                          diffusion_model: BernoulliDiffusion) -> BernoulliDiffusion:
+                          diffusion_model: BernoulliDiffusionModel) -> BernoulliDiffusionModel:
     
     optimizer = optim.SGD(diffusion_model.parameters(), lr=cfg.lr)
