+++ main.py
@@ -1,8 +1,8 @@
 import torch
 
-from config import load_config
-from model import BernoulliDiffusion, ReverseModel
-from train import train_diffusion_model
+from BernoulliDiffusion.config import load_config
+from BernoulliDiffusion.model import BernoulliDiffusionModel, ReverseModel
+from BernoulliDiffusion.train import train_diffusion_model
 
 
 device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
@@ -12,7 +12,7 @@ if __name__ == '__main__':
     cfg = load_config('config.yaml')
 
     reverse_model = ReverseModel(cfg.sequence_length, cfg.T).to(device)
-    diffusion_model = BernoulliDiffusion(reverse_model, cfg.sequence_length, cfg.num_sample_steps, cfg.T).to(device)
+    diffusion_model = BernoulliDiffusionModel(reverse_model, cfg.sequence_length, cfg.num_sample_steps, cfg.T).to(device)
 
     print('Batch of samples from reverse trajectory before training:')
     pre_sample = diffusion_model.p_sample(cfg.batch_size)
